Fuse clint walrus-operator rule into main linter pass
The walrus rule ran a dedicated ast.NodeVisitor over every FunctionDef,
re-walking each function subtree after the main linter had already walked
it. Hook the per-statement-list check into the main Linter's existing
visitors instead, so each block body is scanned once.
